requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
aom==3.3.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
blosc==1.21.1
brotli-bin==1.0.9
brotli==1.0.9
brunsli==0.1
bzip2==1.0.8
c-ares==1.18.1
c-blosc2==2.1.1
ca-certificates==2022.4.26
cached-property==1.5.2
cached_property==1.5.2
cachetools==5.0.0
catboost==1.0.5
certifi==2021.10.8
cffi==1.15.0
cfitsio==4.1.0
charls==2.3.4
charset-normalizer==2.0.12
cloudpickle==2.0.0
cycler==0.11.0
cytoolz==0.11.2
dask-core==2022.5.0
facenet-pytorch==2.5.2
ffmpeg==4.4.0
flatbuffers==2.0
fonttools==4.33.3
freetype==2.10.4
frozenlist==1.3.0
fsspec==2022.3.0
gast==0.5.3
gettext==0.19.8.1
giflib==5.2.1
gmp==6.2.1
gnutls==3.6.13
google-auth-oauthlib==0.4.6
google-auth==2.6.6
google-pasta==0.2.0
grpcio==1.45.0
h11==0.12.0
h5py==3.6.0
hdf5==1.12.1
httpcore==0.14.7
httpx==0.22.0
icu==70.1
idna==3.3
imagecodecs==2022.2.22
imageio==2.19.1
importlib-metadata==4.11.3
jbig==2.1
jpeg==9e
jxrlib==1.1
keras-preprocessing==1.1.2
keras==2.8.0
kiwisolver==1.4.2
krb5==1.19.3
lame==3.100
lcms2==2.12
lerc==3.0
libaec==1.0.6
libavif==0.10.1
libblas==3.9.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libclang==14.0.1
libcurl==7.83.0
libcxx==13.0.1
libdeflate==1.10
libedit==3.1.20191231
libev==4.33
libffi==3.4.2
libgfortran5==11.0.1.dev0
libgfortran==5.0.0.dev0
libiconv==1.16
liblapack==3.9.0
libnghttp2==1.47.0
libopenblas==0.3.20
libpng==1.6.37
libprotobuf==3.20.1
libssh2==1.10.0
libtiff==4.3.0
libvpx==1.11.0
libwebp-base==1.2.2
libwebp==1.2.2
libxcb==1.13
libxml2==2.9.14
libzlib==1.2.11
libzopfli==1.0.3
llvm-openmp==14.0.3
locket==1.0.0
lz4-c==1.9.3
markdown==3.3.6
matplotlib-base==3.5.2
matplotlib==3.5.1
multidict==6.0.2
munkres==1.1.4
ncurses==6.3
nettle==3.6
networkx==2.8
ninja==1.10.2
numpy==1.22.3
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
openh264==2.1.1
openjpeg==2.4.0
openssl==3.0.3
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
partd==1.2.0
pillow==9.1.0
pip==22.0.4
plotly==5.8.0
protobuf==3.20.1
pthread-stubs==0.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pyparsing==3.0.8
python-dateutil==2.8.2
python-graphviz==0.20
python==3.8.13
python_abi==3.8
pytorch-cpu==1.11.0
pytorch==1.11.0
pytz==2022.1
pywavelets==1.3.0
pyyaml==6.0
readline==8.1
requests-oauthlib==1.3.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
setuptools==62.1.0
six==1.15.0
sleef==3.5.1
snappy==1.1.9
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlite==3.38.3
tenacity==8.0.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
tensorflow-deps==2.8.0
tensorflow-macos==2.8.0
tensorflow-metal==0.4.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.5.4
tk==8.6.12
toolz==0.11.2
torchvision==0.9.1
tqdm==4.64.0
typing-extensions==4.2.0
typing_extensions==4.2.0
unicodedata2==14.0.0
urllib3==1.26.9
werkzeug==2.1.2
wheel==0.37.1
wrapt==1.14.1
x264==1!161.3030
x265==3.5
xorg-libxau==1.0.9
xorg-libxdmcp==1.1.3
xz==5.2.5
yaml==0.2.5
yarl==1.7.2
zfp==0.5.5
zipp==3.8.0
zlib-ng==2.0.6
zlib==1.2.11
zstd==1.5.2
